+++ src/index.js
@@ -12,6 +12,8 @@ const { addVote } = require('./services/trainer.js');
 const { stageNames } = require('./config/stageConfig.js');
 const { poll, formatMoney } = require('./utils/utils.js');
 const axios = require('axios');
+const { Flags } = require('discord.js/src/util/BitField.js');
+const { startSpawning, addGuild } = require('./services/spawn.js');
 
 const corsOptions = {
     origin: true,
@@ -36,11 +38,11 @@ client.on(Events.MessageCreate, async (message) => {
 
     runMessageCommand(message, client).then(() => {
         // do nothing
-    }).catch((error) => {
+    }).catch(async (error) => {
         try {
             logger.error(`Error in message command: ${message.content}`);
             logger.error(error);
-            message.reply("There was an error trying to execute that command!");
+            await message.reply("There was an error trying to execute that command!");
         } catch (error) {
             logger.error(`Error in message command: ${message.content}`);
             logger.error(error);
@@ -95,6 +97,7 @@ client.on(Events.InteractionCreate, async (interaction) => {
 // On guild join, log it
 client.on(Events.GuildCreate, (guild) => {
     logger.info(`Joined guild: ${guild.name} (${guild.id})`);
+    addGuild(client, guild);
 });
 
 // When the client is ready, run this code (only once)
@@ -108,6 +111,9 @@ client.once(Events.ClientReady, c => {
         guildString += `${guild.name} (${guild.id})\n`;
     });
     logger.info(`Connected to guilds: \n${guildString}`);
+
+    // start spawners
+    startSpawning(client);
 
     // connect to discordbotlist.com
     if (process.env.STAGE === stageNames.BETA || process.env.STAGE === stageNames.PROD) {